requirements.txt
httpx[http2]
beautifulsoup4
python-telegram-bot>=20.0
APScheduler
lxml
